requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2021.10.8
click==8.1.8
cloudinary==1.29.0
dj-database-url==0.5.0
Django==5.1.7
django-cors-headers==4.7.0
django-on-heroku==1.1.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.115.12
gunicorn==20.1.0
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.3
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.6
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
starlette==0.46.1
typing_extensions==4.12.2
urllib3==1.26.8
uvicorn==0.34.0
whitenoise==6.0.0
